Add device ID validation on client and server using custom proto field option
A new custom option, called regex, is added to the device_id field in the
.proto. The value of the option is then fetched from the proto descriptor
at runtime and used to validate the device ID in the Measurement instance.

An arguably better implementation would be to make protoc plugins to generate
the validation code. To be honest though I didn't want to go to the trouble
of doing that. This implementation is fine for the purpose of this application.

Test: Publish a Measurement using the Pub/Sub logger with a valid device ID
      and validation enabled on both client and server. Publish a Measurement
      with an invalid device ID with validation enabled on the client, and then
      with validation disabled on the client to make sure the server validation
      works.