Make the plot responsive
Do this by setting the width and height of the SVG to 100% in the CSS and
getting rid of explicit width and height attrs on the element. The width is
then defined by the width of the containing element, which is a Bootstrap
column, so changing the Bootstrap column class changes the plot width. Then
set a viewBox on the SVG element and use the viewBox dimensions for laying
out the plots. The viewBox width and height are kind of abstract in the sense
that they form a virtual canvas and the visible width and height now auto-scales
based on the container of the SVG element, so we get resizing on window resize
for free.