Add new field upload_timestamp to Measurement protobuf
upload_timestamp should be set only if upload of a measurement initially
fails and it is stored locally and then uploaded later. This commit
just adds support for this field; it does not set the value. That will
come in a follow-up commit.

Also add a corresponding field to StorableMeasurement and instruct
datastore to omit it if it's empty (i.e. if it's the zero value of
time.Time).