Create ~/.iotcorelogger on init and use it to store anything the program needs to persist
We also create ~/.iotcorelogger/pending and use it for pending
measurements, instead of getting the path to use from a flag.

We use github.com/mitchellh/go-homedir to get the current user's home
directory because it does not require cgo and therefore works when
cross-compiling for ARM, whereas os/user requires cgo.