Add a gRPC service for fetching devices and measurements
The gRPC service, MeasurementService, is defined in the proto file
and cmd/api implements the service. cmd/apiclient is a CLI tool for
calling the service.

measurementservice_config.yaml is the Cloud Endpoints configuration file
for MeasurementService. Cloud Endpoints requires the proto descriptor,
so the protoc command in the Makefile now makes it and it's checked in
as measurementpb/measurement.pb.descriptor.

Finally, this commit adds a Dockerfile that builds and serves the gRPC
server implementing MeasurementService. This container is run alongside
the Extensible Service Proxy (ESP) container for Cloud Endpoints API
management (https://cloud.google.com/endpoints/docs/frameworks/glossary#extensible_service_proxy).