Make String methods generic with tags, proto extensions, and reflection
StorableMeasurement fields now must have a "metric" tag, and optionally
a "unit" tag, to be considered a field that records a measurement value.

Measurement proto fields must have the MeasurementOptions extension set,
with the metric field non-empty. The unit field may be empty.

The tags and extensions are used to format values in the String methods,
for example printing a temperature like this: "temp=18.375°C".