Fix bug where AQI didn't display if it was 0
The AQI value is accessed from a map in the template, and `if $aqi` was
being used to check for its presence. However if it was 0 it would then
be treated as if it wasn't present. Do an actual map membership check
instead. As far as I can tell the only way to do this is with a custom
function. I couldn't find any built-in that does it.